fix(refactor): Move scrolling functions to a separate file
